refactor: extract route groups from server.py into focused router modules (batch 1)
Part of epic #159. Extracts 21 routes into 3 new router files, reducing
server.py from 6,833 to 5,978 lines (-855 lines).

New router modules:
- backend/maxwell_router.py: 8 routes (/api/maxwell/*) + DASHBOARD_FAQ
- backend/assessments_router.py: 8 routes (/api/assessments/*, /api/feature-requests/*, /api/settings/prompt-notes)
- backend/diagnostics_router.py: 5 routes (/api/deployment/git-drift, /api/diagnostics/*, /api/launchers/generate, /api/runner-routing-audit/*)

Each router uses a configure() injection pattern to receive shared helpers
(run_cmd, ORG, REPO_ROOT, etc.) from server.py without circular imports.
All URL paths and HTTP methods are unchanged.
