+++ backend/server.py
@@ -265,9 +265,8 @@ _CACHE_EVICT_BATCH = 50
 # ─── Shared State Locks ───────────────────────────────────────────────────────
 _remediation_history_lock: asyncio.Lock = asyncio.Lock()
 _orchestration_audit_lock: asyncio.Lock = asyncio.Lock()
-_feature_requests_lock: asyncio.Lock = asyncio.Lock()
-_prompt_templates_lock: asyncio.Lock = asyncio.Lock()
-_prompt_notes_lock: asyncio.Lock = asyncio.Lock()
+# _feature_requests_lock, _prompt_templates_lock, _prompt_notes_lock
+# moved to assessments_router.py (epic #159)
 
 # ─── Configuration ────────────────────────────────────────────────────────────
 ORG = os.environ.get("GITHUB_ORG", "D-sorganization")
@@ -5479,283 +5478,16 @@ async def serve_icon():
 # ─── Credentials Probe — see backend/routers/credentials.py ──────────────────
 # Endpoint extracted to routers/credentials.py and registered via app.include_router.
 
-# ─── Maxwell-Daemon endpoints ─────────────────────────────────────────────────
+# ─── Maxwell-Daemon endpoints — see backend/maxwell_router.py ────────────────
+# All /api/maxwell/* routes have been extracted to maxwell_router.py.
+# The router is imported and configured below.
+import maxwell_router as _maxwell_router  # noqa: E402
 
-DASHBOARD_FAQ: dict[str, str] = {
-    "fleet": "The Fleet tab shows all runners in your fleet. Use it to start/stop runners and see hardware metrics.",
-    "remediation": (
-        "The Remediation tab lets you dispatch AI agents (Jules, Codex, Claude) to fix failing CI."
-        " Move to top: Manual Dispatch is the primary control."
-    ),
-    "workflows": (
-        "The Workflows tab lists all GitHub Actions workflows across repos."
-        " Click a workflow to see run history and dispatch it manually."
-    ),
-    "credentials": (
-        "The Credentials tab shows provider connection state."
-        " No secrets are shown - only whether tools are installed and authenticated."
-    ),
-    "assessments": (
-        "The Assessments tab lets you trigger code quality assessments for any repo and view score history."
-    ),
-    "feature-requests": (
-        "The Feature Requests tab dispatches AI agents to implement new features"
-        " with standards injection (TDD, DbC, DRY, LoD)."
-    ),
-    "maxwell": ("The Maxwell tab shows Maxwell-Daemon status and lets you start/stop the service with confirmation."),
-    "queue": "The Queue tab shows live queued and in-progress workflows with auto-refresh every 15 seconds.",
-    "history": "The History tab shows recent workflow runs across all repos, filterable by status.",
-    "machines": "The Machines tab shows hardware telemetry for each fleet node.",
-    "stats": "The Stats tab shows P50/P95 duration analytics and success rates across workflows.",
-    "runner-plan": "The Runner Plan tab manages day/night runner capacity scheduling.",
-    "dispatch": (
-        "To dispatch a remediation agent: go to Remediation tab, select a failed run,"
-        " choose a provider, preview the plan, then dispatch."
-    ),
-    "provider": (
-        "Providers are AI agents: Jules API (cloud, Google), Codex CLI (OpenAI),"
-        " Claude Code CLI (Anthropic), Ollama (local)."
-    ),
-    "loop guard": (
-        "Loop guard prevents infinite retry loops. When the same failure repeats more than"
-        " max_same_failure_attempts times, dispatch is blocked."
-    ),
-}
+# DASHBOARD_FAQ lives in maxwell_router; import here for /api/help/chat below.
+from maxwell_router import DASHBOARD_FAQ  # noqa: E402
 
-
-@app.get("/api/maxwell/status")
-async def get_maxwell_status() -> dict:
-    """Probe Maxwell-Daemon status and connectivity."""
-    import shutil
-
-    maxwell_binary = shutil.which("maxwell") or shutil.which("maxwell-daemon")
-    maxwell_url = os.environ.get("MAXWELL_URL", "")
-    maxwell_port = int(os.environ.get("MAXWELL_PORT", 8322))
-
-    # Check if maxwell service is running via systemd
-    service_running = False
-    service_detail = "unknown"
-    try:
-        import subprocess
-
-        r = subprocess.run(
-            ["systemctl", "is-active", "maxwell-daemon"],
-            capture_output=True,
-            text=True,
-            timeout=5,
-            env=safe_subprocess_env(),
-        )
-        if r.returncode == 0 and r.stdout.strip() == "active":
-            service_running = True
-            service_detail = "systemd service active"
-        else:
-            service_detail = r.stdout.strip() or "not active"
-    except Exception:  # noqa: BLE001
-        service_detail = "systemd probe failed"
-
-    # Check HTTP reachability
-    http_reachable = False
-    http_detail = ""
-    base_url = maxwell_url or f"http://localhost:{maxwell_port}"
-    try:
-        import httpx
-
-        async with httpx.AsyncClient(timeout=3.0) as client:
-            resp = await client.get(f"{base_url}/api/health")
-            http_reachable = resp.status_code == 200
-            http_detail = f"HTTP {resp.status_code}"
-    except Exception as e:  # noqa: BLE001
-        http_detail = str(e)[:80]
-
-    status = "running" if (service_running or http_reachable) else "stopped"
-
-    return {
-        "status": status,
-        "binary_found": maxwell_binary is not None,
-        "binary_path": maxwell_binary,
-        "service_running": service_running,
-        "service_detail": service_detail,
-        "http_reachable": http_reachable,
-        "http_detail": http_detail,
-        "dashboard_url": base_url,
-        "deep_links": {
-            "dashboard": base_url,
-            "health": f"{base_url}/api/health",
-            "logs": "journalctl -u maxwell-daemon -f",
-        },
-        "probed_at": datetime.now(UTC).isoformat(),
-    }
-
-
-@app.post("/api/maxwell/control")
-async def maxwell_control(
-    request: Request,
-    *,
-    principal: Principal = Depends(require_scope("maxwell.control")),  # noqa: B008
-) -> dict:
-    """Start or stop Maxwell-Daemon service (confirmation required)."""
-    body = await request.json()
-    action = str(body.get("action", "")).strip()
-    approved_by = str(body.get("approved_by", "")).strip()
-    if action not in ("start", "stop", "restart"):
-        raise HTTPException(status_code=422, detail="action must be start, stop, or restart")
-    if not approved_by:
-        raise HTTPException(status_code=422, detail="approved_by required for privileged action")
-
-    code, out, stderr = await run_cmd(["systemctl", action, "maxwell-daemon"], timeout=15, cwd=REPO_ROOT)
-    log.info(
-        "maxwell_control: action=%s approved_by=%s exit_code=%d",
-        sanitize_log_value(action),
-        sanitize_log_value(approved_by),
-        code,
-    )
-    if code != 0:
-        log.warning("maxwell %s failed: %s", action, stderr.strip()[:200])
-        raise HTTPException(
-            status_code=502,
-            detail=f"maxwell {action} failed",
-        )
-    return {"status": action + "ed", "action": action, "approved_by": approved_by}
-
-
-# ─── Maxwell-Daemon Proxy Routes ──────────────────────────────────────────────
-
-
-def _maxwell_base_url() -> str:
-    """Return the Maxwell-Daemon base URL from env."""
-    return os.environ.get("MAXWELL_URL", "") or f"http://localhost:{int(os.environ.get('MAXWELL_PORT', 8322))}"
-
-
-@app.get("/api/maxwell/version")
-async def get_maxwell_version() -> dict:
-    """Proxy GET /api/version from Maxwell-Daemon."""
-    path = "/api/version"
-    resp = None
-    try:
-        async with httpx.AsyncClient(timeout=3.0) as client:
-            resp = await client.get(f"{_maxwell_base_url()}{path}")
-            log.info("maxwell_proxy: path=%s status=%s", path, resp.status_code)
-            return resp.json()
-    except Exception as e:  # noqa: BLE001
-        log.info("maxwell_proxy: path=%s status=%s", path, "error")
-        return {"error": str(e)[:120], "daemon_available": False}
-
-        log.info("maxwell_proxy: path=%s status=%s", path, "error")
-        return {"error": str(e)[:120], "daemon_available": False}
-
-
-@app.get("/api/maxwell/daemon-status")
-async def get_maxwell_daemon_status_detail() -> dict:
-    """Proxy GET /api/status from Maxwell-Daemon (pipeline state)."""
-    path = "/api/status"
-    resp = None
-    try:
-        async with httpx.AsyncClient(timeout=3.0) as client:
-            resp = await client.get(f"{_maxwell_base_url()}{path}")
-            log.info("maxwell_proxy: path=%s status=%s", path, resp.status_code)
-            return resp.json()
-    except Exception as e:  # noqa: BLE001
-        log.info("maxwell_proxy: path=%s status=%s", path, "error")
-        return {"error": str(e)[:120], "daemon_available": False}
-
-        log.info("maxwell_proxy: path=%s status=%s", path, "error")
-        return {"error": str(e)[:120], "daemon_available": False}
-
-
-@app.get("/api/maxwell/tasks")
-async def get_maxwell_tasks(limit: int = 20, cursor: str | None = None) -> dict:
-    """Proxy GET /api/tasks from Maxwell-Daemon."""
-    path = "/api/tasks"
-    resp = None
-    try:
-        params: dict = {"limit": limit}
-        if cursor is not None:
-            params["cursor"] = cursor
-        async with httpx.AsyncClient(timeout=3.0) as client:
-            resp = await client.get(f"{_maxwell_base_url()}{path}", params=params)
-            log.info("maxwell_proxy: path=%s status=%s", path, resp.status_code)
-            return resp.json()
-    except Exception as e:  # noqa: BLE001
-        log.info("maxwell_proxy: path=%s status=%s", path, "error")
-        return {"error": str(e)[:120], "daemon_available": False}
-
-        log.info("maxwell_proxy: path=%s status=%s", path, "error")
-        return {"error": str(e)[:120], "daemon_available": False}
-
-
-@app.get("/api/maxwell/tasks/{task_id}")
-async def get_maxwell_task_detail(task_id: str) -> dict:
-    """Proxy GET /api/tasks/{task_id} from Maxwell-Daemon."""
-    path = f"/api/tasks/{task_id}"
-    resp = None
-    try:
-        async with httpx.AsyncClient(timeout=3.0) as client:
-            resp = await client.get(f"{_maxwell_base_url()}{path}")
-            log.info("maxwell_proxy: path=%s status=%s", path, resp.status_code)
-            return resp.json()
-    except Exception as e:  # noqa: BLE001
-        log.info("maxwell_proxy: path=%s status=%s", path, "error")
-        return {"error": str(e)[:120], "daemon_available": False}
-
-        log.info("maxwell_proxy: path=%s status=%s", path, "error")
-        return {"error": str(e)[:120], "daemon_available": False}
-
-
-@app.post("/api/maxwell/dispatch")
-async def maxwell_dispatch_task(
-    request: Request,
-    *,
-    principal: Principal = Depends(require_scope("maxwell.control")),  # noqa: B008
-) -> dict:
-    """Proxy POST /api/dispatch to Maxwell-Daemon (forwards body as-is)."""
-    path = "/api/dispatch"
-    resp = None
-    try:
-        body = await request.body()
-        async with httpx.AsyncClient(timeout=3.0) as client:
-            resp = await client.post(
-                f"{_maxwell_base_url()}{path}",
-                content=body,
-                headers={"Content-Type": "application/json"},
-            )
-            log.info("maxwell_proxy: path=%s status=%s", path, resp.status_code)
-            return resp.json()
-    except Exception as e:  # noqa: BLE001
-        log.info("maxwell_proxy: path=%s status=%s", path, "error")
-        return {"error": str(e)[:120], "daemon_available": False}
-
-        log.info("maxwell_proxy: path=%s status=%s", path, "error")
-        return {"error": str(e)[:120], "daemon_available": False}
-
-
-@app.post("/api/maxwell/pipeline-control/{action}")
-async def maxwell_pipeline_control(
-    action: str,
-    request: Request,
-    *,
-    principal: Principal = Depends(require_scope("maxwell.control")),  # noqa: B008
-) -> dict:
-    """Proxy POST /api/control/{action} to Maxwell-Daemon."""
-    if action not in ("pause", "resume", "abort"):
-        raise HTTPException(status_code=422, detail="action must be pause, resume, or abort")
-    path = f"/api/control/{action}"
-    resp = None
-    try:
-        body = await request.body()
-        async with httpx.AsyncClient(timeout=3.0) as client:
-            resp = await client.post(
-                f"{_maxwell_base_url()}{path}",
-                content=body,
-                headers={"Content-Type": "application/json"},
-            )
-            log.info("maxwell_proxy: path=%s status=%s", path, resp.status_code)
-            return resp.json()
-    except Exception as e:  # noqa: BLE001
-        log.info("maxwell_proxy: path=%s status=%s", path, "error")
-        return {"error": str(e)[:120], "daemon_available": False}
-
-        log.info("maxwell_proxy: path=%s status=%s", path, "error")
-        return {"error": str(e)[:120], "daemon_available": False}
+app.include_router(_maxwell_router.router)
+_maxwell_router.configure(run_cmd, REPO_ROOT)
 
 
 @app.post("/api/help/chat")
@@ -5828,321 +5560,13 @@ async def help_chat(request: Request, *, principal: Principal = Depends(require_
     }
 
 
-# ─── Assessments ──────────────────────────────────────────────────────────────
+# ─── Assessments + Feature Requests — see backend/assessments_router.py ──────
+# All /api/assessments/*, /api/feature-requests/*, and /api/settings/prompt-notes
+# routes have been extracted to assessments_router.py.
+import assessments_router as _assessments_router  # noqa: E402
 
-
-@app.get("/api/assessments/scores")
-async def get_assessment_scores() -> dict:
-    """Return assessment score history from local assessments directory."""
-    assessments_dir = REPO_ROOT / "assessments"
-    results: list[dict] = []
-    if assessments_dir.exists():
-        for score_file in sorted(assessments_dir.rglob("*.json"), reverse=True)[:50]:
-            try:
-                data = json.loads(score_file.read_text(encoding="utf-8"))
-                results.append(
-                    {
-                        "file": str(score_file.relative_to(REPO_ROOT)),
-                        "repo": data.get("repository") or score_file.parent.name,
-                        "score": data.get("score") or data.get("overall_score"),
-                        "date": data.get("date") or data.get("timestamp") or score_file.stat().st_mtime,
-                        "summary": data.get("summary") or data.get("description", "")[:200],
-                        "provider": data.get("provider") or data.get("agent", ""),
-                    }
-                )
-            except Exception:  # noqa: BLE001
-                pass
-    return {"scores": results, "total": len(results)}
-
-
-@app.post("/api/assessments/dispatch")
-async def dispatch_assessment(
-    request: Request,
-    *,
-    principal: Principal = Depends(require_scope("assessments.dispatch")),  # noqa: B008
-) -> dict:
-    """Dispatch an assessment workflow for a repository."""
-    body = await request.json()
-    repo = str(body.get("repository", "")).strip()
-    provider = str(body.get("provider", "jules_api")).strip()
-    ref = str(body.get("ref", "main")).strip()
-    if not repo:
-        raise HTTPException(status_code=422, detail="repository required")
-    if not provider:
-        raise HTTPException(status_code=422, detail="provider required")
-
-    log.info(
-        "audit: assessments_dispatch repo=%s provider=%s ref=%s",
-        sanitize_log_value(repo),
-        sanitize_log_value(provider),
-        sanitize_log_value(ref),
-    )
-
-    # Try to dispatch via GitHub Actions assessment workflow
-    endpoint = f"/repos/{ORG}/Repository_Management/actions/workflows/Jules-Assess-Repo.yml/dispatches"
-    payload = {
-        "ref": "main",
-        "inputs": {"target_repository": f"{ORG}/{repo}", "provider": provider},
-    }
-    with tempfile.NamedTemporaryFile(mode="w", encoding="utf-8", suffix=".json", delete=False) as f:
-        json.dump(payload, f)
-        pf = f.name
-    try:
-        code, _, stderr = await run_cmd(
-            ["gh", "api", endpoint, "--method", "POST", "--input", pf],
-            timeout=30,
-            cwd=REPO_ROOT,
-        )
-    finally:
-        with contextlib.suppress(OSError):
-            Path(pf).unlink()
-    if code != 0:
-        log.warning("assessment dispatch failed: repo=%s stderr=%s", repo, stderr.strip()[:300])
-        raise HTTPException(
-            status_code=502,
-            detail="Assessment dispatch failed",
-        )
-    return {"status": "dispatched", "repository": repo, "provider": provider}
-
-
-# ─── Feature Requests ─────────────────────────────────────────────────────────
-
-_FEATURE_REQUESTS_PATH = Path.home() / "actions-runners" / "dashboard" / "feature_requests.json"
-_PROMPT_TEMPLATES_PATH = Path.home() / "actions-runners" / "dashboard" / "prompt_templates.json"
-_PROMPT_NOTES_PATH = Path.home() / "actions-runners" / "dashboard" / "prompt_notes.json"
-
-STANDARDS_INJECTION: dict[str, str] = {
-    "tdd": (
-        "Use Test-Driven Development: write failing tests first (RED), then minimal code to pass (GREEN),"
-        " then refactor. Tests must pass before any PR."
-    ),
-    "dbc": (
-        "Apply Design by Contract: validate inputs at boundaries, assert internal invariants,"
-        " document pre/postconditions in docstrings."
-    ),
-    "dry": (
-        "Apply DRY: extract shared logic into modules, eliminate duplication."
-        " Three similar code blocks should become one shared function."
-    ),
-    "lod": (
-        "Apply Law of Demeter: components talk to immediate neighbors only."
-        " UI receives view models, not raw nested payloads."
-    ),
-    "security": (
-        "Apply security-first: validate all inputs, avoid injection vulnerabilities,"
-        " use parameterized queries, never log secrets."
-    ),
-    "docs": (
-        "Document public APIs, non-obvious decisions, and architecture choices."
-        " Prefer short clear docstrings over multi-paragraph ones."
-    ),
-}
-
-
-@app.get("/api/feature-requests")
-async def list_feature_requests() -> dict:
-    """List saved feature implementation requests."""
-    try:
-        if _FEATURE_REQUESTS_PATH.exists():
-            data = json.loads(_FEATURE_REQUESTS_PATH.read_text(encoding="utf-8"))
-        else:
-            data = []
-    except Exception:  # noqa: BLE001
-        data = []
-    return {"requests": list(reversed(data[-100:])), "total": len(data)}
-
-
-@app.get("/api/feature-requests/templates")
-async def list_prompt_templates() -> dict:
-    """List saved prompt templates and global prompt notes."""
-    templates_data = []
-    try:
-        if _PROMPT_TEMPLATES_PATH.exists():
-            templates_data = json.loads(_PROMPT_TEMPLATES_PATH.read_text(encoding="utf-8"))
-    except Exception:  # noqa: BLE001
-        pass
-
-    prompt_notes_data = {"notes": "", "enabled": True}
-    try:
-        if _PROMPT_NOTES_PATH.exists():
-            prompt_notes_data = json.loads(_PROMPT_NOTES_PATH.read_text(encoding="utf-8"))
-    except Exception:  # noqa: BLE001
-        pass
-
-    return {
-        "templates": templates_data,
-        "standards": STANDARDS_INJECTION,
-        "promptNotes": prompt_notes_data,
-    }
-
-
-@app.post("/api/feature-requests/templates")
-async def save_prompt_template(
-    request: Request,
-    *,
-    principal: Principal = Depends(require_scope("feature-requests.manage")),  # noqa: B008
-) -> dict:
-    """Save a prompt template."""
-    body = await request.json()
-    name = str(body.get("name", "")).strip()
-    content = str(body.get("content", "")).strip()
-    if not name or not content:
-        raise HTTPException(status_code=422, detail="name and content required")
-    async with _prompt_templates_lock:
-        try:
-            templates: list[dict] = []
-            if _PROMPT_TEMPLATES_PATH.exists():
-                templates = json.loads(_PROMPT_TEMPLATES_PATH.read_text(encoding="utf-8"))
-            existing_idx = next((i for i, t in enumerate(templates) if t.get("name") == name), None)
-            template = {
-                "name": name,
-                "content": content,
-                "updated_at": datetime.now(UTC).isoformat(),
-            }
-            if existing_idx is not None:
-                templates[existing_idx] = template
-            else:
-                templates.append(template)
-            config_schema.atomic_write_json(_PROMPT_TEMPLATES_PATH, templates)
-        except Exception as e:  # noqa: BLE001
-            raise HTTPException(status_code=500, detail=str(e)) from e
-    return {"status": "saved", "name": name}
-
-
-@app.get("/api/settings/prompt-notes")
-async def get_prompt_notes() -> dict:
-    """Get the global prompt notes that are automatically injected into every prompt."""
-    try:
-        if _PROMPT_NOTES_PATH.exists():
-            data = json.loads(_PROMPT_NOTES_PATH.read_text(encoding="utf-8"))
-        else:
-            data = {"notes": "", "enabled": True}
-    except Exception:  # noqa: BLE001
-        data = {"notes": "", "enabled": True}
-    return data
-
-
-@app.put("/api/settings/prompt-notes")
-async def update_prompt_notes(request: Request, *, principal: Principal = Depends(require_scope("operator"))) -> dict:  # noqa: B008
-    """Update the global prompt notes."""
-    body = await request.json()
-    if not isinstance(body, dict):
-        raise HTTPException(status_code=422, detail="expected object body")
-
-    notes = str(body.get("notes", "")).strip()
-    enabled = bool(body.get("enabled", True))
-
-    async with _prompt_notes_lock:
-        try:
-            data = {"notes": notes, "enabled": enabled}
-            config_schema.atomic_write_json(_PROMPT_NOTES_PATH, data)
-        except Exception as e:  # noqa: BLE001
-            raise HTTPException(status_code=500, detail=str(e)) from e
-    return {"status": "saved", "notes_length": len(notes), "enabled": enabled}
-
-
-@app.post("/api/feature-requests/dispatch")
-async def dispatch_feature_request(
-    request: Request,
-    *,
-    principal: Principal = Depends(require_scope("feature-requests.manage")),  # noqa: B008
-) -> dict:
-    """Dispatch a feature implementation request via CI remediation workflow."""
-    client_ip = request.client.host if request.client else "unknown"
-    check_dispatch_rate(client_ip)
-    body = await request.json()
-    repo = str(body.get("repository", "")).strip()
-    branch = str(body.get("branch", "main")).strip()
-    provider = str(body.get("provider", "jules_api")).strip()
-    prompt = str(body.get("prompt", "")).strip()
-    standards = body.get("standards", []) or []
-    template_id = str(body.get("template_id", "")).strip()
-    if not repo:
-        raise HTTPException(status_code=422, detail="repository required")
-    if not prompt and not template_id:
-        raise HTTPException(status_code=422, detail="prompt or template_id required")
-
-    log.info(
-        "audit: feature_request_dispatch repo=%s provider=%s branch=%s",
-        sanitize_log_value(repo),
-        sanitize_log_value(provider),
-        sanitize_log_value(branch),
-    )
-
-    # Load and apply prompt notes if enabled
-    prompt_notes_data: dict[str, object] = {"notes": "", "enabled": True}
-    try:
-        if _PROMPT_NOTES_PATH.exists():
-            prompt_notes_data = json.loads(_PROMPT_NOTES_PATH.read_text(encoding="utf-8"))
-    except Exception:  # noqa: BLE001
-        pass
-
-    # Build full prompt with notes and standards injection
-    full_prompt = prompt
-    notes_val = str(prompt_notes_data.get("notes", ""))
-    if prompt_notes_data.get("enabled", True) and notes_val.strip():
-        full_prompt = f"{notes_val}\n\n{prompt}"
-
-    injected_standards = "\n\n".join(
-        f"[{s.upper()}] {STANDARDS_INJECTION[s]}" for s in standards if s in STANDARDS_INJECTION
-    )
-    if injected_standards:
-        full_prompt = f"{full_prompt}\n\n## Engineering Standards\n{injected_standards}"
-
-    # Save to history
-    entry: dict = {}
-    async with _feature_requests_lock:
-        try:
-            history: list[dict] = []
-            if _FEATURE_REQUESTS_PATH.exists():
-                history = json.loads(_FEATURE_REQUESTS_PATH.read_text(encoding="utf-8"))
-            entry = {
-                "id": str(int(datetime.now(UTC).timestamp())),
-                "repository": repo,
-                "branch": branch,
-                "provider": provider,
-                "prompt": prompt[:500],
-                "standards": list(standards),
-                "status": "dispatched",
-                "created_at": datetime.now(UTC).isoformat(),
-            }
-            history.append(entry)
-            config_schema.atomic_write_json(_FEATURE_REQUESTS_PATH, history[-200:])
-        except Exception:  # noqa: BLE001
-            pass
-
-    # Dispatch via feature-request workflow
-    endpoint = f"/repos/{ORG}/Repository_Management/actions/workflows/Jules-Feature-Request.yml/dispatches"
-    payload = {
-        "ref": "main",
-        "inputs": {
-            "target_repository": f"{ORG}/{repo}",
-            "branch": branch,
-            "provider": provider,
-            "prompt": full_prompt[:10000],
-        },
-    }
-    with tempfile.NamedTemporaryFile(mode="w", encoding="utf-8", suffix=".json", delete=False) as f:
-        json.dump(payload, f)
-        pf = f.name
-    try:
-        code, _, stderr = await run_cmd(
-            ["gh", "api", endpoint, "--method", "POST", "--input", pf],
-            timeout=30,
-            cwd=REPO_ROOT,
-        )
-    finally:
-        with contextlib.suppress(OSError):
-            Path(pf).unlink()
-    if code != 0:
-        log.warning("feature_request_dispatch failed: %s", stderr.strip()[:200])
-        # Don't raise - save history record and return success anyway (workflow may not exist yet)
-    return {
-        "status": "dispatched",
-        "repository": repo,
-        "provider": provider,
-        "entry_id": entry.get("id", ""),
-    }
+app.include_router(_assessments_router.router)
+_assessments_router.configure(run_cmd, ORG, REPO_ROOT, check_dispatch_rate)
 
 
 # ─── Fleet Orchestration Control Plane ───────────────────────────────────────
@@ -6516,297 +5940,18 @@ async def fleet_orchestration_deploy(
     }
 
 
-# ─── Diagnostics & Launchers ──────────────────────────────────────────────────
+# ─── Diagnostics, Launchers & Runner-Routing-Audit — see diagnostics_router.py ─
+# All /api/deployment/git-drift, /api/diagnostics/*, /api/launchers/generate, and
+# /api/runner-routing-audit/* routes extracted to diagnostics_router.py.
+import diagnostics_router as _diagnostics_router  # noqa: E402
 
-
-@app.get("/api/deployment/git-drift")
-async def get_git_drift() -> dict:
-    """Return git-commit-based drift: compares HEAD against origin/main."""
-    repo_root = Path(__file__).parent.parent
-    result: dict[str, object] = {}
-
-    source = "unknown"
-    try:
-        out = subprocess.run(
-            ["git", "rev-parse", "HEAD"],
-            capture_output=True,
-            text=True,
-            timeout=5,
-            cwd=repo_root,
-        )
-        source = out.stdout.strip()
-        result["source_commit"] = source[:12]
-    except Exception:  # noqa: BLE001
-        result["source_commit"] = "unknown"
-
-    try:
-        out = subprocess.run(
-            ["git", "rev-parse", "origin/main"],
-            capture_output=True,
-            text=True,
-            timeout=5,
-            cwd=repo_root,
-        )
-        remote = out.stdout.strip()
-        result["remote_commit"] = remote[:12]
-        result["is_drifted"] = bool(source and remote and source != remote)
-        if result["is_drifted"]:
-            result["drift_details"] = "deployed version differs from origin/main"
-        else:
-            result["drift_details"] = "up to date"
-    except Exception:  # noqa: BLE001
-        result["is_drifted"] = False
-        result["remote_commit"] = "unknown"
-        result["drift_details"] = "could not reach origin/main"
-
-    result["process_pid"] = os.getpid()
-    return result
-
-
-@app.get("/api/diagnostics/summary")
-async def get_diagnostics_summary() -> dict:
-    """Consolidated diagnostics for the Diagnostics tab."""
-    summary: dict[str, object] = {}
-
-    # WSL status
-    try:
-        wsl_result = subprocess.run(
-            ["wsl", "-l", "-v"],
-            capture_output=True,
-            text=True,
-            timeout=10,
-            encoding="utf-16-le",
-            errors="replace",
-        )
-        summary["wsl_status"] = wsl_result.stdout.strip()
-        summary["wsl_available"] = wsl_result.returncode == 0
-    except Exception:  # noqa: BLE001
-        try:
-            wsl_result_raw = subprocess.run(
-                ["wsl", "-l", "-v"],
-                capture_output=True,
-                timeout=10,
-            )
-            summary["wsl_status"] = wsl_result_raw.stdout.decode("utf-16-le", errors="replace").strip()
-            summary["wsl_available"] = wsl_result_raw.returncode == 0
-        except Exception:  # noqa: BLE001
-            summary["wsl_status"] = "WSL not available"
-            summary["wsl_available"] = False
-
-    # Dashboard process info
-    proc = psutil.Process(os.getpid())
-    summary["dashboard_pid"] = proc.pid
-    summary["dashboard_memory_mb"] = round(proc.memory_info().rss / 1024 / 1024, 1)
-    summary["dashboard_port"] = PORT
-
-    # Git commit
-    try:
-        out = subprocess.run(
-            ["git", "rev-parse", "--short", "HEAD"],
-            capture_output=True,
-            text=True,
-            timeout=5,
-            cwd=Path(__file__).parent.parent,
-        )
-        summary["git_commit"] = out.stdout.strip() or "unknown"
-    except Exception:  # noqa: BLE001
-        summary["git_commit"] = "unknown"
-
-    # Drift info
-    try:
-        drift = await get_git_drift()
-        summary["is_drifted"] = drift.get("is_drifted", False)
-        summary["source_commit"] = drift.get("source_commit", "unknown")
-        summary["remote_commit"] = drift.get("remote_commit", "unknown")
-        summary["drift_details"] = drift.get("drift_details", "")
-    except Exception:  # noqa: BLE001
-        summary["is_drifted"] = False
-
-    return summary
-
-
-@app.post("/api/diagnostics/restart-service")
-async def restart_dashboard_service(
-    request: Request,
-    *,
-    principal: Principal = Depends(require_scope("system.control")),  # noqa: B008
-) -> dict:
-    """Restart the dashboard systemd service (WSL/Linux only, localhost only)."""
-    client = request.client
-    if not client or client.host not in ("127.0.0.1", "::1"):
-        raise HTTPException(status_code=403, detail="Local access only")
-
-    try:
-        result = subprocess.run(
-            [SYSTEMCTL_BIN, "--user", "restart", "runner-dashboard"],
-            capture_output=True,
-            text=True,
-            timeout=30,
-        )
-        return {
-            "success": result.returncode == 0,
-            "output": (result.stdout + result.stderr).strip(),
-        }
-    except Exception as exc:  # noqa: BLE001
-        log.exception("Failed to restart runner-dashboard service")
-        raise HTTPException(status_code=500, detail="Restart failed") from exc
-
-        log.exception("Failed to restart runner-dashboard service")
-        raise HTTPException(status_code=500, detail="Restart failed") from exc
-
-
-@app.post("/api/launchers/generate")
-async def generate_launchers(
-    request: Request,
-    principal: Principal = Depends(require_scope("system.control")),  # noqa: B008
-) -> dict:
-    """Generate Windows PowerShell launcher scripts on the Desktop."""
-    output_dir = Path.home() / "Desktop" / "RunnerDashboard"
-    output_dir.mkdir(parents=True, exist_ok=True)
-
-    launchers_created: list[str] = []
-
-    script = output_dir / "Open-Dashboard.ps1"
-    script.write_text('Start-Process "http://localhost:8321"\n', encoding="utf-8")
-    launchers_created.append(str(script))
-
-    keepalive = output_dir / "Start-WSL-Keepalive.ps1"
-    keepalive.write_text(
-        'Start-ScheduledTask -TaskName "WSL-Dashboard-Keepalive" -ErrorAction SilentlyContinue\n'
-        'Write-Host "Keepalive task started"\n',
-        encoding="utf-8",
-    )
-    launchers_created.append(str(keepalive))
-
-    restart = output_dir / "Restart-Dashboard-Service.ps1"
-    restart.write_text(
-        'wsl -e bash -c "systemctl --user restart runner-dashboard && echo Service restarted"\n',
-        encoding="utf-8",
-    )
-    launchers_created.append(str(restart))
-
-    diag = output_dir / "Open-Diagnostics.ps1"
-    diag.write_text('Start-Process "http://localhost:8321/#diagnostics"\n', encoding="utf-8")
-    launchers_created.append(str(diag))
-
-    log.info("Generated %d launcher scripts in %s", len(launchers_created), output_dir)
-    return {
-        "output_dir": str(output_dir),
-        "launchers": launchers_created,
-        "message": f"Created {len(launchers_created)} launcher scripts in {output_dir}",
-    }
-
-
-# ─── Hosted-Runner Billing Audit ─────────────────────────────────────────────
-
-HOSTED_RUNNER_PATTERNS = re.compile(
-    r"^(ubuntu-|windows-|macos-|GitHub Actions \d|Hosted Agent)",
-    re.IGNORECASE,
-)
-_runner_audit_cache: dict[str, Any] = {"violations": [], "last_checked": None, "error": None}
-_runner_audit_lock = asyncio.Lock()
-
-
-@app.get("/api/runner-routing-audit")
-async def get_runner_routing_audit() -> JSONResponse:
-    """Return recent workflow runs that executed on GitHub-hosted runners."""
-    return JSONResponse(_runner_audit_cache)
-
-
-@app.post("/api/runner-routing-audit/refresh")
-async def refresh_runner_routing_audit() -> JSONResponse:
-    """Trigger an immediate audit refresh."""
-    asyncio.create_task(_run_runner_audit())
-    return JSONResponse({"status": "refresh triggered"})
-
-
-async def _run_runner_audit() -> None:
-    async with _runner_audit_lock:
-        try:
-            violations = await _fetch_hosted_runner_violations()
-            _runner_audit_cache["violations"] = violations
-            _runner_audit_cache["last_checked"] = datetime.now(UTC).isoformat().replace("+00:00", "Z")
-            _runner_audit_cache["error"] = None
-            if violations:
-                log.warning(
-                    "BILLING ALERT: %d workflow run(s) detected on GitHub-hosted runners",
-                    len(violations),
-                )
-        except Exception as exc:  # noqa: BLE001
-            _runner_audit_cache["error"] = str(exc)
-            log.error("Runner routing audit failed: %s", exc)
-
-
-async def _fetch_hosted_runner_violations() -> list[dict[str, Any]]:
-    """Query GitHub API for recent runs on hosted runners across org repos."""
-    token = os.environ.get("GH_TOKEN", "").strip()
-    if not token:
-        return []
-
-    headers = {"Authorization": f"Bearer {token}", "Accept": "application/vnd.github+json"}
-    org = ORG
-    violations: list[dict[str, Any]] = []
-
-    async with httpx.AsyncClient(timeout=30) as client:
-        repos_resp = await client.get(
-            f"https://api.github.com/orgs/{org}/repos",
-            headers=headers,
-            params={"per_page": 50, "sort": "pushed"},
-        )
-        if repos_resp.status_code != 200:
-            return []
-        repos = [r["name"] for r in repos_resp.json()]
-
-        for repo_name in repos[:20]:  # limit to 20 most recently pushed
-            try:
-                runs_resp = await client.get(
-                    f"https://api.github.com/repos/{org}/{repo_name}/actions/runs",
-                    headers=headers,
-                    params={"per_page": 10, "status": "completed"},
-                )
-                if runs_resp.status_code != 200:
-                    continue
-                for run in runs_resp.json().get("workflow_runs", []):
-                    jobs_resp = await client.get(
-                        f"https://api.github.com/repos/{org}/{repo_name}/actions/runs/{run['id']}/jobs",
-                        headers=headers,
-                        params={"per_page": 30},
-                    )
-                    if jobs_resp.status_code != 200:
-                        continue
-                    for job in jobs_resp.json().get("jobs", []):
-                        runner_name = job.get("runner_name") or ""
-                        runner_group = job.get("runner_group_name") or ""
-                        if HOSTED_RUNNER_PATTERNS.match(runner_name) or runner_group == "GitHub Actions":
-                            violations.append(
-                                {
-                                    "repo": repo_name,
-                                    "workflow": run.get("name", ""),
-                                    "run_id": run["id"],
-                                    "job_name": job.get("name", ""),
-                                    "runner_name": runner_name,
-                                    "runner_group": runner_group,
-                                    "run_url": run.get("html_url", ""),
-                                    "started_at": job.get("started_at", ""),
-                                    "conclusion": job.get("conclusion", ""),
-                                }
-                            )
-            except Exception:  # noqa: BLE001
-                continue
-
-    return violations
-
-
-async def _runner_audit_loop() -> None:
-    await asyncio.sleep(30)  # initial delay
-    while True:
-        await _run_runner_audit()
-        await asyncio.sleep(900)  # 15 minutes
+app.include_router(_diagnostics_router.router)
+_diagnostics_router.configure(PORT, SYSTEMCTL_BIN, ORG, REPO_ROOT)
 
 
 @app.on_event("startup")
 async def _start_background_tasks() -> None:
-    asyncio.create_task(_runner_audit_loop())
+    asyncio.create_task(_diagnostics_router.runner_audit_loop())
 
 
 # ─── Main ─────────────────────────────────────────────────────────────────────
